Fix type of argument of public header

+++ Sources/CsevenZip/include/sevenzip.h
@@ -5,17 +5,17 @@
 #include "../7zCrc.h"
 
 // #define SzArEx_IsDir(p, i) (SzBitArray_Check((p)->IsDirs, i))
-int SevenZip_SzArEx_IsDir(const CSzArEx *p, size_t fileIndex) {
+int SevenZip_SzArEx_IsDir(const CSzArEx *p, UInt32 fileIndex) {
     return SzArEx_IsDir(p, fileIndex);
 }
 
 // #define SzArEx_GetFileSize(p, i) ((p)->UnpackPositions[(i) + 1] - (p)->UnpackPositions[i])
-UInt64 SevenZip_SzArEx_GetFileSize(const CSzArEx *p, size_t fileIndex) {
+UInt64 SevenZip_SzArEx_GetFileSize(const CSzArEx *p, UInt32 fileIndex) {
     return SzArEx_GetFileSize(p, fileIndex);
 }
 
 // #define SzBitWithVals_Check(p, i) ((p)->Defs && ((p)->Defs[(i) >> 3] & (0x80 >> ((i) & 7))) != 0)
-int SevenZip_SzBitWithVals_Check(const CSzBitUi64s *p, size_t i) {
+int SevenZip_SzBitWithVals_Check(const CSzBitUi64s *p, UInt32 i) {
     return SzBitWithVals_Check(p, i);
 }
